Added: Add UriScheme and move UriUtils to com.termux.shared.net package
