+++ app/src/main/java/com/termux/app/TermuxService.java
@@ -26,7 +26,7 @@ import com.termux.app.settings.properties.TermuxAppSharedProperties;
 import com.termux.app.terminal.TermuxTerminalSessionClient;
 import com.termux.app.utils.PluginUtils;
 import com.termux.shared.data.IntentUtils;
-import com.termux.shared.data.UriUtils;
+import com.termux.shared.net.UriUtils;
 import com.termux.shared.models.errors.Errno;
 import com.termux.shared.shell.ShellUtils;
 import com.termux.shared.termux.shell.TermuxShellEnvironmentClient;
