+++ termux-shared/src/main/java/com/termux/shared/net/UriUtils.java
@@ -1,25 +1,26 @@
-package com.termux.shared.data;
+package com.termux.shared.net;
 
-import android.content.ContentResolver;
 import android.net.Uri;
 
 import androidx.annotation.NonNull;
 
+import com.termux.shared.data.DataUtils;
 import com.termux.shared.file.FileUtils;
+import com.termux.shared.models.net.UriScheme;
 
 public class UriUtils {
 
     /**
      * Get the full file path from a {@link Uri}.
      *
-     *  If the {@link Uri} was created from file path with {@link Uri#parse(String)}, like "am"
-     *  command "-d" option does, and the path contained a "#", then anything after it would become
-     *  the fragment and {@link Uri#getPath()} will only return the path before it, which would be
-     *  invalid. The fragment must be manually appended to the path to get the full path.
+     * If the {@link Uri} was created from file path with {@link Uri#parse(String)}, like "am"
+     * command "-d" option does, and the path contained a "#", then anything after it would become
+     * the fragment and {@link Uri#getPath()} will only return the path before it, which would be
+     * invalid. The fragment must be manually appended to the path to get the full path.
      *
-     *  If the {@link Uri} was created with {@link Uri.Builder} and path was set
-     *  with {@link Uri.Builder#path(String)}, then "#" will automatically be encoded to "%23"
-     *  and separate fragment will not exist.
+     * If the {@link Uri} was created with {@link Uri.Builder} and path was set
+     * with {@link Uri.Builder#path(String)}, then "#" will automatically be encoded to "%23"
+     * and separate fragment will not exist.
      *
      * @param uri The {@link Uri} to get basename from.
      * @return Returns the file path if found, otherwise {@code null}.
@@ -55,45 +56,45 @@ public class UriUtils {
     }
 
     /**
-     * Get {@link ContentResolver#SCHEME_FILE} {@link Uri} for path.
+     * Get {@link UriScheme#SCHEME_FILE} {@link Uri} for path.
      *
      * @param path The path for the {@link Uri}.
      * @return Returns the {@link Uri}.
      */
     public static Uri getFileUri(@NonNull String path) {
-        return new Uri.Builder().scheme(ContentResolver.SCHEME_FILE).path(path).build();
+        return new Uri.Builder().scheme(UriScheme.SCHEME_FILE).path(path).build();
     }
 
     /**
-     * Get {@link ContentResolver#SCHEME_FILE} {@link Uri} for path.
+     * Get {@link UriScheme#SCHEME_FILE} {@link Uri} for path.
      *
      * @param authority The authority for the {@link Uri}.
      * @param path The path for the {@link Uri}.
      * @return Returns the {@link Uri}.
      */
     public static Uri getFileUri(@NonNull String authority, @NonNull String path) {
-        return new Uri.Builder().scheme(ContentResolver.SCHEME_FILE).authority(authority).path(path).build();
+        return new Uri.Builder().scheme(UriScheme.SCHEME_FILE).authority(authority).path(path).build();
     }
 
     /**
-     * Get {@link ContentResolver#SCHEME_CONTENT} {@link Uri} for path.
+     * Get {@link UriScheme#SCHEME_CONTENT} {@link Uri} for path.
      *
      * @param path The path for the {@link Uri}.
      * @return Returns the {@link Uri}.
      */
     public static Uri getContentUri(@NonNull String path) {
-        return new Uri.Builder().scheme(ContentResolver.SCHEME_CONTENT).path(path).build();
+        return new Uri.Builder().scheme(UriScheme.SCHEME_CONTENT).path(path).build();
     }
 
     /**
-     * Get {@link ContentResolver#SCHEME_CONTENT} {@link Uri} for path.
+     * Get {@link UriScheme#SCHEME_CONTENT} {@link Uri} for path.
      *
      * @param authority The authority for the {@link Uri}.
      * @param path The path for the {@link Uri}.
      * @return Returns the {@link Uri}.
      */
     public static Uri getContentUri(@NonNull String authority, @NonNull String path) {
-        return new Uri.Builder().scheme(ContentResolver.SCHEME_CONTENT).authority(authority).path(path).build();
+        return new Uri.Builder().scheme(UriScheme.SCHEME_CONTENT).authority(authority).path(path).build();
     }
 
 }
